Update to add an instance_id attribute.
zookeeper V3.5.2-alpha dynamically generates a configuration file and updates the configuration file generated by chef to move the ensemble information to the dynamic file.   In my chef cookbook, I will generate the dynamic file  and stop this cookbook from generating a config file that zookeeper will update.  I will do this by not setting the ensemble field. That works but the problem is with my_id.